# backend/donald/rest/openapi.yaml
...
openapi: 3.1.0
info:
    title: Donald
    version: 1.0.0
servers:
    - url: http://localhost:8080/v1
paths:
...
    /agent-edges/{id}:
        get:
            tags:
                - AgentEdge
            summary: Get AgentEdge by ID
            operationId: getAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentEdge
            summary: Partially update AgentEdge
            operationId: patchAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEdge'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEdge'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentEdge
            summary: Delete AgentEdge by ID
            operationId: deleteAgentEdge
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /agent-events/{id}:
        get:
            tags:
                - AgentEvent
            summary: Get AgentEvent by ID
            operationId: getAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentEvent
            summary: Partially update AgentEvent
            operationId: patchAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentEvent'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentEvent'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentEvent
            summary: Delete AgentEvent by ID
            operationId: deleteAgentEvent
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /agent-runs/{id}:
        get:
            tags:
                - AgentRun
            summary: Get AgentRun by ID
            operationId: getAgentRun
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRun'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - AgentRun
            summary: Partially update AgentRun
            operationId: patchAgentRun
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/AgentRun'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/AgentRun'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - AgentRun
            summary: Delete AgentRun by ID
            operationId: deleteAgentRun
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
    /clients/{id}:
        get:
            tags:
                - Client
            summary: Get Client by ID
            operationId: getClient
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "200":
                    description: Successful operation
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Client'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        patch:
            tags:
                - Client
            summary: Partially update Client
            operationId: patchClient
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/Client'
                required: true
            responses:
                "200":
                    description: Successful update
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/Client'
                "400":
                    description: Bad Request
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Version conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
        delete:
            tags:
                - Client
            summary: Delete Client by ID
            operationId: deleteClient
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                    type: string
            responses:
                "204":
                    description: No Content
                "404":
                    description: Not Found
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "409":
                    description: Conflict
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
                "500":
                    description: Internal Server Error
                    content:
                        application/problem+json:
                            schema:
                                $ref: '#/components/schemas/Problem'
...
components:
    schemas:
        AgentEdge:
            properties:
                condition_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                edge_type:
                    $ref: '#/components/schemas/AgentEdgeType'
                from_node_uuid:
                    format: uuid
                    type: string
                run_uuid:
                    format: uuid
                    type: string
                status:
                    $ref: '#/components/schemas/AgentEdgeStatus'
                to_node_uuid:
                    format: uuid
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - from_node_uuid
                - to_node_uuid
                - edge_type
                - status
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
...
        AgentEdgeStatus:
            enum:
                - invalid
                - pending
                - traversed
                - skipped
                - removed
            type: string
        AgentEdgeType:
            enum:
                - invalid
                - dependency
                - transition
                - conditional
                - retry
                - fallback
                - subagent_spawn
            type: string
        AgentEvent:
            properties:
                agent_label:
                    nullable: true
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                event_type:
                    $ref: '#/components/schemas/AgentEventType'
                idempotency_key:
                    type: string
                node_uuid:
                    format: uuid
                    nullable: true
                    type: string
                occurred_at:
                    format: date-time
                    type: string
                payload:
                    $ref: '#/components/schemas/AgentEventPayload'
                    nullable: true
                run_uuid:
                    format: uuid
                    type: string
                sequence:
                    format: int64
                    type: integer
                status:
                    $ref: '#/components/schemas/AgentEventStatus'
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - run_uuid
                - sequence
                - event_type
                - idempotency_key
                - occurred_at
                - status
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
...
        AgentEventStatus:
            enum:
                - invalid
                - recorded
                - discarded
            type: string
        AgentEventType:
            enum:
                - invalid
                - run_started
                - run_updated
                - run_finished
                - plan_declared
                - node_added
                - node_updated
                - node_status_changed
                - node_removed
                - edge_added
                - edge_updated
                - edge_removed
                - artifact_added
                - intervention_requested
                - intervention_delivered
                - intervention_resolved
                - heartbeat
                - agent_message
            type: string
...
        AgentRun:
            properties:
                agent_identifier:
                    nullable: true
                    type: string
                agent_summary:
                    nullable: true
                    type: string
                client_uuid:
                    format: uuid
                    type: string
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                display_summary:
                    nullable: true
                    type: string
                finished_at:
                    format: date-time
                    nullable: true
                    type: string
                graph_revision:
                    format: int64
                    type: integer
                initial_prompt:
                    nullable: true
                    type: string
                last_event_sequence:
                    format: int64
                    type: integer
                last_heartbeat_at:
                    format: date-time
                    nullable: true
                    type: string
                name:
                    nullable: true
                    type: string
                protocol_version:
                    nullable: true
                    type: string
                run_key:
                    type: string
                started_at:
                    format: date-time
                    nullable: true
                    type: string
                status:
                    $ref: '#/components/schemas/AgentRunStatus'
                status_message:
                    nullable: true
                    type: string
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - client_uuid
                - run_key
                - status
                - graph_revision
                - last_event_sequence
                - created_by
                - updated_by
                - created_at
                - updated_at
            type: object
...
        AgentRunStatus:
            enum:
                - invalid
                - in_progress
                - failed
                - blocked_on_user_decision
                - blocked_on_missing_data
                - blocked_on_provider_outage
                - succeeded
                - pending
                - cancelled
            type: string
...
        Client:
            properties:
                created_at:
                    format: date-time
                    type: string
                created_by:
                    format: uuid
                    type: string
                industry:
                    $ref: '#/components/schemas/ClientIndustry'
                    nullable: true
                name:
                    type: string
                status:
                    $ref: '#/components/schemas/ClientStatus'
                updated_at:
                    format: date-time
                    type: string
                updated_by:
                    format: uuid
                    type: string
                uuid:
                    format: uuid
                    type: string
            required:
                - uuid
                - name
                - created_by
                - updated_by
                - status
                - created_at
                - updated_at
            type: object
        ClientIndustry:
            enum:
                - invalid
                - fintech
                - furniture
            type: string
...
        ClientStatus:
            enum:
                - invalid
                - active
                - inactive
            type: string
...
        Problem:
            properties:
                detail:
                    type: string
                instance:
                    type: string
                status:
                    type: integer
                title:
                    type: string
                type:
                    type: string
            type: object