# backend/donald/rest/server/delete_agent_run.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/gofrs/uuid"

	"github.com/nextwave/donald/core/module/agent_run/types"
)

func (h *handlers) DeleteAgentRun(w http.ResponseWriter, r *http.Request) {

	UUIDStr := chi.URLParam(r, "id")
	UUIDVal, err := uuid.FromString(UUIDStr)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid path parameter UUID: %w", err))
		return
	}

	if delErr := h.core.AgentRun().Delete(r.Context(), types.DeleteRequest{
		UUID: UUIDVal,
	}); delErr != nil {
		if delErr.Error() == "entity not found" {
			writeProblem(w, http.StatusNotFound, delErr)
			return
		}
		writeProblem(w, http.StatusInternalServerError, delErr)
		return
	}

	w.WriteHeader(http.StatusNoContent)
}