# backend/donald/rest/server/get_agent_event.go
// Code generated by nuzur go-code-gen. DO NOT EDIT.

package server

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/gofrs/uuid"

	"github.com/nextwave/donald/core/module/agent_event/types"
)

func (h *handlers) GetAgentEvent(w http.ResponseWriter, r *http.Request) {

	UUIDStr := chi.URLParam(r, "id")
	UUIDVal, err := uuid.FromString(UUIDStr)
	if err != nil {
		writeProblem(w, http.StatusBadRequest, fmt.Errorf("invalid path parameter UUID: %w", err))
		return
	}

	fetchRes, err := h.core.AgentEvent().FetchAgentEventByUUID(r.Context(), types.FetchAgentEventByUUIDRequest{
		UUID: UUIDVal,
	})
	if err != nil {
		writeProblem(w, http.StatusInternalServerError, err)
		return
	}

	if len(fetchRes.Results) == 0 {
		writeProblem(w, http.StatusNotFound, fmt.Errorf("entity not found"))
		return
	}

	writeJSON(w, http.StatusOK, fetchRes.Results[0])
}